[file-order] Vocabulary class changes to VocabularyManager
